cleanup: Remove redundant clones
Summary:
Suggested by `cargo clippy`. This is the last phab in the stack; after this, `cargo clippy` is clean (at least for version `clippy 0.1.57 (f1edd042 2021-11-29)`).

Many of the ones on `Arc<App>` are obviously more intentional than others. I think this check is useful in general for detecting unintentional other clones, so I went and removed those `clone`s anyways, even though we get some statements like `let cloned_app = app;`. The borrow checker will catch us if we try to double-move a given `Arc<App>`, and `cargo clippy` will catch us if we leave an extra `clone` in, so we should be able to statically confirm that we have the runtime-optimal number of `clone`s actually in the codebase.

Example:

```
warning: redundant clone
  --> focus/internals/src/lib/util/sandbox_command.rs:86:29
   |
86 |         let cloned_app = app.clone();
   |                             ^^^^^^^^ help: remove this
   |
note: this value is dropped without further use
  --> focus/internals/src/lib/util/sandbox_command.rs:86:26
   |
86 |         let cloned_app = app.clone();
   |                          ^^^
   = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#redundant_clone

```

Test Plan:
```
$ cargo test
$ cargo clippy  # no more diagnostics
```

Reviewers: wilhelm

Differential Revision: https://phabricator.twitter.biz/D805276