cleanup: Remove dead #[allow(unused)] annotations
Summary: Some of these annotations are dead. Where they're not dead, I replaced them with `#[allow(dead_code)]`, since that appears to be the type of warning which is actually surfaced. Using the same annotation for dead code should make it easier to go through these sites later if we want to re-vivify them.

Test Plan:
```
$ cargo test
```

Reviewers: wilhelm

Differential Revision: https://phabricator.twitter.biz/D805343