[focus] Add `.arcconfig`
Summary:
This makes it so that `arc multi-diff` knows which branch to diff against by default. (The global default set in `/etc/arcconfig` is `origin/master`, but this repository uses `origin/main`.)

Obviously, we won't need this once we move into source.

Test Plan: Run `arc multi-diff` and don't get prompted to select a branch to diff against.

Reviewers: wilhelm

Differential Revision: https://phabricator.twitter.biz/D822054