git-maintenance runner
The git-maintenance scheduled runner uses for-each-repo, which has the
unfortunate behavior of exiting early if any of the repo tasks fails. For
example, if you have repos A, B, and C, and while running `git maintenance` in
B, the prefetch operation fails, maintenance tasks will never be run in C.

This runner performs the same functionality as for-each-repo, but doesn't bail
if it encounters an error. IF one of the mainternance tasks fails for a given
repo, it will log the error and continue to the next.

As a feature, the runner will automatically prune entries in the
`maintenance.repos` key if it discovers that one of the repos in the list is
no longer on disk.

Lastly, there's a convenience function that will set up the default mainteance
config options for a repo.

JIRA Issues: PFD-137

Differential Revision: https://phabricator.twitter.biz/D838167