Add model supporting the outlining concept that can write sparse checkout patterns
The new model will be used to write sparse-checkout files and eventually supplant all of the special-case code that is strewn throughout sparse_repos, etc.

---
Do not remove this text section from the summary. It is required
to create the dependency. You may remove it after the revision is
created.
Depends on D848297

Differential Revision: https://phabricator.twitter.biz/D848314