Sparse repo sync
Rewrite a lot of Focus to support operation without a dense repository.

This review is big. I don't know how much more I can break it down given the time constraints. If you have suggestions, fire away.

A change of note is that we should have operation::clone use operation::init, but as of now, operation::clone presents a better UX since it doesn't require a clone over the network. We should add that functionality to operation::init and remove the cloning-specific code from operation::clone.

Differential Revision: https://phabricator.twitter.biz/D848542