[focus] Add `Label` type
Summary:
This type lets us parse the components of a Bazel target up-front and use the data later. Currently, Pants labels are similar, but we've just left them as `String`s.

It might be a little over-engineered; we can simplify it according to reviewer suggestions, if any.

Test Plan:
Added unit tests for parsing.

```
$ cargo test
```

Differential Revision: https://phabricator.twitter.biz/D850108