[focus] Update API for `ObjectDatabase` trait
Summary:
As discussed with @dbernadett, the object database might want to use the original key value as part of lookups/insertions, so change the interface to accept them.

This is kind of clumsy since we return the resulting `ContentHash` via a tuple, so I'm open to an alternative interface.

Test Plan:
```
$ cargo test
```

Reviewers: #focus, dbernadett

Subscribers: dbernadett

JIRA Issues: PFD-143

Differential Revision: https://phabricator.twitter.biz/D854978