[focus][easy] Add comment about hashing external packages
Summary: In principle, we could hash the contents of external packages. However, the version of those external packages is determined strictly by the contents of the `WORKSPACE` file (and any of its `load`ed `.bzl` files), so we can skip content hashing of external packages and rely on the implicit dependency on `WORKSPACE` instead.

Test Plan:
```
$ cargo test
```

Reviewers: #focus

JIRA Issues: PFD-145

Differential Revision: https://phabricator.twitter.biz/D863393