[focus] Pass `ObjectDatabase` into `Repo::sync`
Summary: Right now, it's hard-coded to use the `SimpleGitOdb`, which means that it doesn't work if you're trying to use the performant RocksDB object database.

Test Plan:
Run

```
$ cargo run -- -C ~/workspace/source-focus2 index clear
$ cargo run -- -C ~/workspace/source-focus2 index resolve 'bazel://strato/bin/...'
```

Before this commit, it produces this warning:

```
Keys STILL missing, this is a bug:
760060d06f1ac3fecdd92b3c785fafa0caa7e16e bazel://strato/bin
```

After this commit, it succeeds without producing the above warning.

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D875931