[focus] Query top-level targets explicitly in Bazel resolver
Summary: We need to examine the top-level targets of the specified focused packages more carefully than we were doing before. We need to determine which of them are Bazel-compatible, which could be either because they have the `bazel-compatible` tag, or because they were defined in a `BUILD.bazel` file.

Test Plan:
Added test.

```
$ cargo test --workspace
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D875930