[focus] Don't double-query `deps` during Bazel resolution
Summary:
For a given package, we only need one level of dependency information up-front. Our content-hashing mechanism traverses the dependency chain later to assign a content-addressable hash to a given package (one which embeds dependency information).

The ideal flow looks like this:

- Find all packages we need;
- query for all their dependencies at once;
- query for all of the package + dependency information at once and assemble the data in memory.

This requires just two Bazel queries.

But right now, it actually looks like this:

- Find all the packages we need;
- query for all their dependencies at once;
- then, for each package or dependency:
  - query for its *transitive* dependencies and assemble the data in memory.

This requres O(n) Bazel queries, which is really slow in practice.

This commit gets us to an intermediate point:

- Find all the packages we need;
- query for all their dependencies at once;
- then, for each package or dependency:
  - query for its *immediate* dependencies and assemble the data in memory.

which is not asymptotically faster, but in practice saves us a lot of time/work.

This just requires a few adjustments to collect all of the appropriate intermediate dependencies.

Test Plan:
Existing dependency graph tests.

```
$ cargo test
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D881644