[focus][easy] Add missing feature to `focus-internals`
Summary:
When compiling this package in isolation, it fails:

```
$ cargo build -p focus-internals
error[E0433]: failed to resolve: could not find `ArgEnum` in `clap`
  --> focus/internals/src/lib/operation/index.rs:21:11
   |
21 |     clap::ArgEnum,
   |           ^^^^^^^ could not find `ArgEnum` in `clap`

For more information about this error, try `rustc --explain E0433`.
error: could not compile `focus-internals` due to previous error
```

But when compiling them all together, it succeeds:

```
$ cargo build --workspace
```

This is because of Rust's feature-flag accumulation. It finds all the features required by any package in the build graph and then enables them for the entire compilation. So this issue only appears when building this package in isolation.

Test Plan:
```
$ for i in focus-commands focus-internals focus-tracing focus-testing focus-tracing focus-util tool-insights-client distributed-memoization; do cargo build -p "$i"; done
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D880115