[focus][easy] Remove extra `@` for external packages
Summary: The comment on `Label.external_package` says that it includes the `@`, so it was wrong to add it again here. In practice, this had little significance because we don't traverse external packages anyways for content hashing or dependency tracking (we rely on the hash of `WORKSPACE` for that), but the messaging was confusing.

Test Plan:
Run

```
$ cargo run -- -C ~/workspace/source-focus2/ index generate
```

and observe that we no longer get reports of missing keys in the form `@@foo/bar`.

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D881202