[focus] Refactor `BazelResolve::run_bazel_query`
Summary: Unify the two implementations (XML output vs newline-separated output). Note that now the behavior is to read the entire string into memory before returning it for processing. I don't think this is going to be a serious issue, but we can change it to return something streaming in the future.

Test Plan:
```
$ cargo test
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D882478