[focus][easy] Mark fixture repo targets as `bazel-compatible`
Summary: Right now, the Bazel resolver needs these targets to be marked as `bazel-compatible`. The fact that this worked at all before was incidental, since source files don't need to be marked as `bazel-compatible`, so we materialized directories with source files. This changes in the next phab.

Test Plan:
```
$ cargo test
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D882480