[focus] Resolve dependencies of the form `//foo/...` directly
Summary: It turns out that `//foo/...` can expand to multiple Bazel packages without `//foo` itself being a Bazel package. To deal with this, we just take note of the fact that a layer requested to get the dependencies for `//foo/...`, and as part of querying the transitive dependencies for `//foo/...`, we add anything under `foo/` to a synthetic dependency key with the name `//foo/...`.

Test Plan:
Added test.

```
$ cargo test
```

This also starts working with only a handful of missing targets (for unrelated reasons), where before we had 100+ missing keys:

```
$ cargo run -- -C ~/workspace/source-focus2 index resolve bazel://easypromote/...
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D882481