[focus] Lift match on `dep_key` in `get_files_to_materialize`
Summary: This pattern was a little bit strange, in which we first got a `DependencyKey`, and then looked up the corresponding `DependencyValue`s, and then looked up *that* `DependencyValue's` corresponding `DependencyKey`s (if any). This can be simplified and made more general by handling both cases of the `DependencyKey`s uniformly.

Test Plan:
```
$ cargo test
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D883826