[focus] Preserve exact package names during Bazel resolution
Summary: Previously, we used `buildfiles(deps(...))` to ensure that we picked up all of the build-file dependencies. However, this means that we drop the information of profile which packages are necessary for the layer, so we end up trying to reconstruct it later by querying Bazel and asking for top-level packages and so on. This ends up requiring a lot of hacks, which we can remove by preserving the package names that we were initially trying to resolve.

Test Plan:
```
$ cargo test
$ cargo run -- -C ~/workspace/source-focus2 index generate  # succeeds!
```

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D883828