[focus] Cache `load` dependencies
Summary: Due to the new way that we handle `//foo/...` dependencies, we need to traverse the entire tree looking for build-graph-relevant-files. In order to minimize the impact of this, cache the parsed `load` dependencies.

Test Plan:
```
$ cargo test
```

`focus sync` time taken drops from 18s to 7s for me, which is roughly what it was before.

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D883827