[focus] Parallelize resolution only by subresolver type
Summary: Currently, we parallelize resolution by querying for every coordinate individually. However, the Bazel resolver is designed to resolve multiple coordinates simultaneously, and we waste a lot of time waiting on `bazel query` running it on every individual coordinate. (There's an internal `Mutex` ensuring that we don't call Bazel multiple times in parallel anyways.) This commit changes it so that we only parallelize subrequests if they would be routed to different resolver types.

Test Plan:
```
$ cargo test
```

And try `focus index generate` and confirm that it completes faster and with many fewer Bazel queries.

Reviewers: #focus

JIRA Issues: PFD-144

Differential Revision: https://phabricator.twitter.biz/D883829