Memoization Synchonization via git
These changes implement and test MemoizationCacheSynchronizer which can synchronize the RocksDBMemoizationCache using a remote git repo as an intermediary. A keyset_id is used as a git-tag name to identify a git tree object that hold key->value pairs. Re-publishing a keyset_id will overwrite the tag, causing new fetches of that keyset_id to return the new tree object. I expect that focus will use source commit_ids as keyset_id's.

When publishing a new keyset, it is possible to hint which existing keyset the new one was derived from. This may aid the implementation of delta compression for these keysets.

Feel confident with the correctness. Feeling less confident with style. Going to work on getting some performance numbers in a different phab.

Another concern with the above code is my use of the App struct. Not sure if I am using it correctly, but I did need it to use the git_helpers.

JIRA Issues: PFD-148

Differential Revision: https://phabricator.twitter.biz/D867651