[ti-client]: Make TI client write logs that are readable to all users
The TI client is currently not writing logs that can be consumed by the tool insights
daemon. This happened when we started using `NamedTempFile` instead of handrolling temp
files. `NamedTempFile` writes the log with the mode `0o100600` which makes it unreadable
to the tool insights daemon with runs as the MDE user.
We're fixing this by changing the mode of the temp file to `0o100644`. The security risk
is similar because all users could list the directory before as well.

JIRA Issues: PFD-192

Differential Revision: https://phabricator.twitter.biz/D886265