For hourly tasks (fetch), schedule every 5 minutes with random offset
To match the trickle-fetch frequency, have the background git-maintenance
"hourly" tasks run every 5 minutes. It would probably make sense to review
the use of these hourly/daily/weekly flags with git-maintenance and instead
just run the tasks we want to run directly, but we'll save that for another
time.

JIRA Issues: PFD-178

Differential Revision: https://phabricator.twitter.biz/D887559