Provision outlining trees with sparse patterns from a file in the repo
Before performing outlining, attempt to apply a sparse profile to the outlining tree from a file stored in the repo at `focus/outlining.patterns.json`. In addition to being more general, this allows for subsequent updates to the patterns applied in the outlining tree, which may be important as repos evolve.

This also resolves an issue where sparse patterns were reapplied when they didn't need to be because of the precedence of Patterns emitted by the code returning items from the dependency cache -- this may have eventually led to flaky tests in addition to being less than optimal.

I also changed the trace console formatter we use during testing from "full" output to "pretty" output because the former was very difficult to read.

JIRA Issues: PFD-194

Differential Revision: https://phabricator.twitter.biz/D892303