[focus] Refactor `WorkingTree` to contain a `git2::Repository`
Summary: We need to do this if we want the `WorkingTree` to be able to return a reference to e.g. a `Commit`, since that `Commit`'s lifetime is bounded by the repository's lifetime. This also identifies a few places where we e.g. create a `git2::Repository`, then get the path out of it and put it in the `WorkingTree`, and then might create the repository again later with `.git_repo()`

Test Plan:
```
$ cargo test --workspace
```

Reviewers: #focus, wilhelm

JIRA Issues: PFD-148

Differential Revision: https://phabricator.twitter.biz/D893779