[focus] Fix non-deterministic content-hashing for external repository packages
Summary:
In the comments here, it says that we don't need to do anything else to construct the hash for a package with an external repository set.

This is technically true from a correctness standpoint, but it causes errors to be raised later in the pipeline when we actually try to store the incoherent content. To fix this, just mix in the full label name into every hash.

I tried combining the `BazelPackage` and `BazelBuildFile` code paths, but it ended up being confusing and I couldn't get the tests to path. The main difference is that the `BazelPackage` hashes the entire directory and any build-relevant files in it, while the `BazelBuildFile` only processes a single `.bzl` file.

Test Plan:
The following now succeeds:

```
$ cargo run --release -- -C ~/workspace/source-focus4 index clear
$ cargo run --release -- -C ~/workspace/source-focus4 sync
```

Reviewers: #focus, wilhelm

JIRA Issues: PFD-148

Differential Revision: https://phabricator.twitter.biz/D895821