[focus][easy] Use variable named `sparse_repo_path` everywhere in `index`
Summary: There was a mistake where we used the wrong repo path for `focus index resolve` (always `.` instead of whatever was provided by the caller). I don't think this affected correctness in the way I happened to invoke it, but it could in the future.

Test Plan:
```
$ cargo test --workspace
```

Reviewers: #focus

JIRA Issues: PFD-201

Differential Revision: https://phabricator.twitter.biz/D895966