[focus][easy] Lower non-package dependency key diagnostic to `debug!`
Summary: This case will trigger if the caller passes e.g. a `DependencyKey::Path`, which is not unreasonable if they resolved multiple kinds of targets and are just passing us the result. For example, they might have resolved `directory:foo/bar`. In those cases, we ignore adding them to the object database, since they can't fall out-of-date. I'm just leaving this in case it's helpful for debugging in the future at some point.

Test Plan:
```
$ cargo check --workspace
```

Reviewers: #focus, wilhelm

Reviewed By: #focus, wilhelm

Subscribers: wilhelm

JIRA Issues: PFD-201

Differential Revision: https://phabricator.twitter.biz/D900992