[focus] Warn only if racy hashes are different
Under the single-threaded model, we expect that the cache will never be populated between the read and write times, so we indiscriminately warned about non-deterministic hashing if we encountered a value having been written to the cache before we expected it to be. Under a multi-threaded optimistic caching model, it's possible for someone else to beat us to writing the cached value, in which case this warning triggers spuriously. We only need to warn if someone wrote a value to the cache which was *different* from the value we just computed. If the value was the same, then it was deterministic.

JIRA Issues: PFD-209

Differential Revision: https://phabricator.twitter.biz/D915580