[focus] Use thread-local dependency cycle detection
Previously, we assumed that there was only a single reader/writer to the caches, so we reused them to determine if there was a dependency cycle by writing a sentinel `Calculating` value to the cache for an entry we were visiting higher up in the call stack. Under an optimistic parallel model, it's possible that two parties will be calculating the same value at the same time, which produced this error erroneously. To fix this, restrict the dependency cycle detection to the same thread.

JIRA Issues: PFD-209

Differential Revision: https://phabricator.twitter.biz/D915589