[focus] Don't recursively hash directory for `/...` dependencies
I mentioned before in chat that currently, for a `/...` dependency, we traverse the entire containing tree and hash any `BUILD` files that we see. This is so that any `load` statements in a sub-package are included in the dependencies (see test). I also mentioned that we could skip this if we made it so that the `/...` package had an actual dependency on each subpackage, instead of "absorbing" their dependencies by reading all of their `BUILD` files.

Well, it turns out that we actually create those dependencies already, so we can just remove the bit of the code that does the recursive dependency traversal. The tough part is just writing the test to validate the assumption that the `/...` dependency will be invalidated correctly.

Content-hashing time dropped to ~2s after this change (from ~6s):

```
2022-06-17T22:37:08.974716Z focus_internals::model::repo: Checking cache for sparse checkout patterns
2022-06-17T22:37:10.948620Z focus_internals::model::repo: Cache hit for sparse checkout patterns num_seen_keys=6661
2022-06-17T22:37:11.007680Z focus_internals::model::repo: Applying patterns profile="/Users/wkhan/workspace/source-focus/.git/info/sparse-checkout" count=1347
```

JIRA Issues: PFD-49, PFD-209

Differential Revision: https://phabricator.twitter.biz/D917741