[focus] Change `try_label_into_path` to `label_into_path`
Summary: This commit is to address the the revert in {D921854}. Previously, it wasn't possible for us to try to materialize a path on disk due to a `load` file on an external dependency. After making the change to respect `prelude_bazel`, this is now possible, because the `prelude_bazel` file can have a `load` of an external repository.

Test Plan:
Undo the revert in {D921854}, then try running `focus sync`.

Before this commit:

```
$ cargo run -- -C ~/workspace/source-focus sync
...
Error: Sync failed

Caused by:
    Cannot read dependency on external repository for label: Label("@rules_pkg///...")
```

After this commit:

```
$ cargo run -- -C ~/workspace/source-focus sync
... succeeds ...
```

Reviewers: #focus, wilhelm, O57517 focus:/

JIRA Issues: PFD-226

Differential Revision: https://phabricator.twitter.biz/D922465