[focus] Create `GitBinary` struct
Summary: The purpose of this struct is to isolate all calls to Git statically and ensure that calls to Git don't have their environment polluted accidentally. For example, tests pass on local machines because you have `user.name`/`user.email` set, but not on remote CI machines. To ensure that the behavior is the same between the two, we explicitly call `env_clear` on `Command`s spawned from the `GitBinary`.

Test Plan:
Used in debugging CI failures on Github.

```
$ cargo test --workspace
```

Reviewers: #focus, jsimms, O57517 focus:/

JIRA Issues: PFD-238

Differential Revision: https://phabricator.twitter.biz/D925891