[focus] Fix missing-key panic in `focus index push`
A panic would occur in `focus index push` if we tried to look up a `DependencyKey::BazelBuildFile`, since we don't actually write those values to the cache. In our test fixtures, we never had a dependency graph which included a `.bzl` file, so this situation was never exercised.

JIRA Issues: PFD-251

Differential Revision: https://phabricator.twitter.biz/D939713