Make commands work in directories below root
Summary:
Commands don't currently work in directories below the root of the focus repo.
This makes it so all commands with a sparse_repo var try to find the nearest (traversing to parent dirs) dir with a `.focus` dir.

Test Plan: cargo test --workspace and testing the CLI manually

Reviewers: dbernadett, O57517 focus:/

JIRA Issues: PFD-267

Differential Revision: https://phabricator.twitter.biz/D947458