build: use `[patch]` for our patched git2
If we take a dependency on a library which depends on upstream git2, then the version resolver will be unable to find a suitable version of git2 to use. To get around this, we use Cargo's `[patch]` functionality to instead replace `git2` where it appears in the build graph with the patched version, without affecting version resolution. Used in the next commit to take a dependency on `vergen`.