refactor(content_hash): create custom `Error` type
When using shared futures (later), the return value is cached so that it's not computed more than once. This implies that callers need to be able to get a copy of the return value when they `.await` the future, so the error type has to be cloneable. However, `anyhow::Error` is not. To fix this, we create our own `Error` type which is.