Add focus pull
Summary:
Focus pull is a new command that will update the refs from prefetch and
update the current branch from `refs/focus/sync`.
The idea is that by using `focus/sync`, the user will always be on a
commit that exists in the dependency index cache and therefore reduce
the `sync` times.

JIRA Issue: PFD-288