Add helpful error message when main branch is outside of shallow_since window.
Summary:
When a user would create a new repo, but the local dense repo had a main branch that would land outside of the shallow_since window, new would fail with an unhelpful error message.

`error no commits selected for shallow requests`

This catches the condition in which this message would be returned from the `git clone` process and instead gives a message that details the problem and how to fix it.

Test Plan:
I reset master in my local dense repo to an old commit in history. (5 months back)

This is the result of `focus new`:
```
test focus.dev new test-too-old-master-on-new
2022-10-13T17:07:32.016947Z focus: Created sandbox path="/Users/jconnerton/Library/Logs/focus/focus_sandbox_new_d77nCH"
2022-10-13T17:07:32.088207Z focus: Cloning "~/workspace/source" into /Users/jconnerton/workspace/test/test-too-old-master-on-new
2022-10-13T17:07:32.089200Z run: focus_operations::clone: Dense repo path: /Users/jconnerton/workspace/source sparse_repo_path="/Users/jconnerton/workspace/test/test-too-old-master-on-new" clone_args=CloneArgs { origin: Some(Local("/Users/jconnerton/workspace/source")), branch: "master", projects_and_targets: [], copy_branches: true, days_of_history: 90, do_post_clone_fetch: true } template=None tracker=Tracker { _tempdir: None, directory: "/Users/jconnerton/Library/Application Support/focus" } app=App
Error: Failed to clone the repository

Caused by:
    Your main branch master is older than the specified shallow date: 2022-07-15. Run a `git pull` in your dense repo to update it!
```

JIRA Issues: PFD-305